# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Sediba Aesthetic & Wellness Clinic
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/initiate:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Create a Yoco payment attempt for a pending booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiatePaymentBody"
      responses:
        "200":
          description: Yoco hosted-checkout redirect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentFormResponse"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Booking not payable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/status:
    get:
      operationId: getPaymentStatus
      tags: [payments]
      summary: Get booking + payment status by booking reference
      parameters:
        - name: ref
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current booking and payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentStatusResponse"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    InitiatePaymentBody:
      type: object
      properties:
        bookingRef:
          type: string
      required:
        - bookingRef
    PaymentFormResponse:
      type: object
      properties:
        url:
          type: string
        mode:
          type: string
          enum: [live, test]
      required:
        - url
        - mode
    PaymentStatusResponse:
      type: object
      properties:
        bookingStatus:
          type: string
        paymentStatus:
          type: string
        amountCents:
          type: integer
        paidCents: { type: integer }
        outstandingCents: { type: integer }
        inSalon: { type: boolean }
        refundedCents: { type: integer }
        netReceiptsCents: { type: integer }
      required:
        - bookingStatus
        - paymentStatus
        - amountCents
        - paidCents
        - outstandingCents
        - inSalon
        - refundedCents
        - netReceiptsCents
...